+++ src/includes/graph.hpp
@@ -5,15 +5,16 @@
 #include <queue>
 #include <limits>
 #include <algorithm>
+#include <cstdint>
 
 /**
  * Represents an edge in the graph.
  */
 struct Edge {
     int target;
-    int weight;
+    uint32_t weight;
     
-    Edge(int target, int weight) : target(target), weight(weight) {}
+    Edge(int target, uint32_t weight) : target(target), weight(weight) {}
 };
 
 /**
@@ -23,7 +24,7 @@ class Graph {
 public:
     Graph();
     void add_edge(int source, int target, int weight);
-    std::vector<int> shortest_path(int source, int target, long long& totalTime);
+    std::vector<int> shortest_path(int source, int target, uint32_t& totalTime);
     int get_max_node_id() const;
 
 private:
